moved AddItem to comp and removed ListSwitch dupe
